+++ src/scripts/add/debt.mjs
@@ -1,5 +1,5 @@
 import { addToDb } from '../db.mjs';
-import { updateBackLink, sanitize } from '../utilities/index.mjs';
+import { updateBackLink, sanitize } from '../helpers/index.mjs';
 
 updateBackLink();
 document.addEventListener('submit', async (event) => {
