+++ src/scripts/insights/monthly.mjs
@@ -1,5 +1,5 @@
 import { getAllFromIndex, getAllFromCloudIndex } from '../db.mjs';
-import { formatCurrency } from '../utilities/index.mjs';
+import { formatCurrency } from '../helpers/index.mjs';
 
 const today = new Date();
 let month = today.getMonth();
